Rename classes to align with replacement of Validator with Calibrator
