+++ src/test/java/com/softwareonpurpose/uinavigator/validators/ClassCalibrator.java
@@ -1,0 +1,35 @@
+package com.softwareonpurpose.uinavigator.validators;
+
+import com.softwareonpurpose.calibrator4test.Calibrator;
+
+public class ClassCalibrator extends Calibrator {
+
+    private static final String description = "Class";
+    private final Class actual;
+    private final Class expected;
+
+    private ClassCalibrator(Class actual, Class expected) {
+        super(description, expected, actual);
+        this.actual = actual;
+        this.expected = expected;
+    }
+
+    public static ClassCalibrator getInstance(Class expected, Class actual) {
+        return new ClassCalibrator(expected, actual);
+    }
+
+    @Override
+    protected void executeVerifications() {
+        verify("Canonical Name", expected.getCanonicalName(), actual.getCanonicalName());
+    }
+
+    @Override
+    protected boolean actualExists() {
+        return actual != null;
+    }
+
+    @Override
+    protected boolean expectedExists() {
+        return expected != null;
+    }
+}
